Moved testing functions to uvdata/utils.py
